Remove expat-package target from 1700+

+++ makefiles/expat.mk
@@ -27,7 +27,9 @@ expat: expat-setup
 	$(call AFTER_BUILD,copy)
 endif
 
+ifeq ($(shell [ "$(CFVER_WHOLE)" -ge 1700 ] && echo 1),1)
 expat-package: expat-stage
+else
 	# expat.mk Package Structure
 	rm -rf $(BUILD_DIST)/{expat,libexpat1{,-dev}}
 	mkdir -p $(BUILD_DIST)/expat/$(MEMO_PREFIX)$(MEMO_SUB_PREFIX)/bin \
@@ -54,5 +56,5 @@ expat-package: expat-stage
 
 	# expat.mk Build cleanup
 	rm -rf $(BUILD_DIST)/{expat,libexpat1{,-dev}}
-
+endif
 .PHONY: expat expat-package
